rename xcurrencyexplorer to xblockexplorer
